refactor(google-drive): extract Drive client into shared helper
PR #8 landed the Drive API logic inline inside FetchGoogleDriveAbility:
HTTP, pagination, native-file export, binary streaming, error mapping,
and URL parsing all lived in one 731-line file. With multiple one-shot
abilities (list, read-doc, download) about to land on top of the same
Drive API surface, that meant copy-pasting all of it three times.

Extract everything HTTP-shaped into a stateless GoogleDriveClient
under Handlers/GoogleDrive/, taking an access token in the constructor
and exposing list_folder(), get_file_metadata(), export_native(),
download_binary(), plus extract_folder_id() / extract_file_id()
helpers and the export-candidates table.

Also adds AuthHelper under Abilities/GoogleDrive/ so each ability gets
a single get_access_token() call that handles provider lookup, scope
check, and token refresh in one shot — surfacing googledrive_scope_missing
the same way across every Drive ability.

FetchGoogleDriveAbility now delegates all Drive API work to the client
and keeps only the pipeline-shaped behavior (binary download into the
flow-scoped uploads dir via DirectoryManager, payload shape for the
fetch handler). Output shape and error codes are byte-for-byte
compatible with PR #8.

No behavior change.
