+++ inc/Abilities/GoogleDrive/FetchGoogleDriveAbility.php
@@ -2,24 +2,18 @@
 /**
  * Fetch Google Drive Ability.
  *
- * Pure business logic — no handler config, no engine data, no pipeline
- * context. Any caller (REST, CLI, chat tool, pipeline handler) can
- * invoke this directly.
+ * Pipeline-shaped Drive fetcher — accepts a download target (pipeline
+ * + flow IDs) so binary files can be streamed to the flow-scoped
+ * uploads directory. Any caller (REST, CLI, chat tool, pipeline
+ * handler) can invoke this directly, but binary downloads only happen
+ * when a flow context is supplied. Callers that want a one-shot
+ * download into the WP uploads area should call the dedicated
+ * `datamachine/download-googledrive` ability instead.
  *
- * Resolves auth via the unified `google` OAuth provider — single
- * credential covers every Google handler in this plugin. See the
- * plugin loader for the scope union.
- *
- * Behavior:
- * - Lists files in a Drive folder via files.list with pagination.
- * - Optionally recurses into subfolders.
- * - Optionally filters by MIME type and modifiedTime.
- * - Native Google Docs / Sheets / Slides are exported to text/csv.
- * - Other binary files are streamed to the Data Machine uploads area
- *   when a download target (pipeline/flow context) is provided.
- * - Google Forms / Sites and other non-exportable native types are
- *   returned with `payload.skipped = true` so the caller can decide
- *   whether to log and continue.
+ * Resolves auth via the unified `google` OAuth provider. All HTTP /
+ * pagination / export / error mapping lives in
+ * `Handlers\GoogleDrive\GoogleDriveClient` — this class is just the
+ * pipeline glue.
  *
  * @package DataMachineBusiness
  * @subpackage Abilities\GoogleDrive
@@ -30,18 +24,12 @@ namespace DataMachineBusiness\Abilities\GoogleDrive;
 
 use DataMachine\Abilities\PermissionHelper;
 use DataMachine\Core\FilesRepository\DirectoryManager;
-use DataMachine\Core\FilesRepository\FilesystemHelper;
+use DataMachineBusiness\Abilities\GoogleDrive\AuthHelper;
+use DataMachineBusiness\Handlers\GoogleDrive\GoogleDriveClient;
 
 defined( 'ABSPATH' ) || exit;
 
 class FetchGoogleDriveAbility {
-
-	private const API_BASE         = 'https://www.googleapis.com/drive/v3';
-	private const LIST_FIELDS      = 'nextPageToken,files(id,name,mimeType,modifiedTime,size,owners,webViewLink,md5Checksum,parents)';
-	private const PAGE_SIZE        = 100;
-	private const REQUIRED_SCOPE   = 'https://www.googleapis.com/auth/drive.readonly';
-	private const MAX_RECURSION    = 25;
-	private const RETRY_AFTER_CAP  = 60;
 
 	private static bool $registered = false;
 
@@ -131,46 +119,24 @@ class FetchGoogleDriveAbility {
 			return $this->fail( 'folder_id is required.', $logs );
 		}
 
-		$recursive       = ! empty( $input['recursive'] );
-		$mime_filter     = isset( $input['mime_filter'] ) && is_array( $input['mime_filter'] ) ? array_values( array_filter( array_map( 'strval', $input['mime_filter'] ) ) ) : array();
-		$modified_since  = isset( $input['modified_since'] ) ? trim( (string) $input['modified_since'] ) : '';
 		$download_target = isset( $input['download_target'] ) && is_array( $input['download_target'] ) ? $input['download_target'] : array();
 
-		$auth_provider = self::get_auth_provider();
-		if ( ! $auth_provider ) {
-			return $this->fail( 'Google authentication provider is not registered.', $logs );
-		}
-
-		if ( ! $auth_provider->has_scope( self::REQUIRED_SCOPE ) ) {
-			return $this->fail(
-				sprintf(
-					/* translators: %s: OAuth scope URL. */
-					'The connected Google account is missing the Drive scope (%s). Disconnect and reconnect the Google integration in Data Machine settings to grant Drive access.',
-					self::REQUIRED_SCOPE
-				),
-				$logs
-			);
-		}
-
-		$access_token = $auth_provider->get_service();
+		$access_token = AuthHelper::get_access_token();
 		if ( is_wp_error( $access_token ) ) {
 			return $this->fail( $access_token->get_error_message(), $logs );
 		}
-		if ( ! is_string( $access_token ) || '' === $access_token ) {
-			return $this->fail( 'Failed to obtain a Google access token.', $logs );
-		}
 
-		$files   = array();
-		$visited = array();
+		$client = new GoogleDriveClient( $access_token );
 
-		$list_result = $this->list_folder_recursive(
+		$list_result = $client->list_folder(
 			$folder_id,
-			$access_token,
-			$recursive,
-			$mime_filter,
-			$modified_since,
-			0,
-			$visited,
+			array(
+				'recursive'      => ! empty( $input['recursive'] ),
+				'mime_filter'    => isset( $input['mime_filter'] ) && is_array( $input['mime_filter'] )
+					? array_values( array_filter( array_map( 'strval', $input['mime_filter'] ) ) )
+					: array(),
+				'modified_since' => isset( $input['modified_since'] ) ? trim( (string) $input['modified_since'] ) : '',
+			),
 			$logs
 		);
 
@@ -178,15 +144,9 @@ class FetchGoogleDriveAbility {
 			return $this->fail( $list_result->get_error_message(), $logs );
 		}
 
+		$files = array();
 		foreach ( $list_result as $raw ) {
-			// Folders are returned by the recursion helper as well so the
-			// caller can introspect structure, but for the fetch handler
-			// we only emit leaf files.
-			if ( $this->is_folder( $raw ) ) {
-				continue;
-			}
-
-			$payload = $this->materialize_file( $raw, $access_token, $download_target, $logs );
+			$payload = $this->materialize_file( $raw, $client, $download_target, $logs );
 
 			$files[] = array(
 				'id'            => $raw['id'] ?? '',
@@ -213,136 +173,23 @@ class FetchGoogleDriveAbility {
 	}
 
 	/**
-	 * Recursively list files in a folder.
-	 *
-	 * @param string   $folder_id      Drive folder ID.
-	 * @param string   $access_token   OAuth bearer token.
-	 * @param bool     $recursive      Whether to recurse into subfolders.
-	 * @param string[] $mime_filter    Optional MIME types to include.
-	 * @param string   $modified_since Optional ISO-8601 timestamp.
-	 * @param int      $depth          Current recursion depth (caps at MAX_RECURSION).
-	 * @param array    $visited        Visited folder IDs to break cycles.
-	 * @param array    $logs           Logs reference.
-	 * @return array|\WP_Error
-	 */
-	private function list_folder_recursive(
-		string $folder_id,
-		string $access_token,
-		bool $recursive,
-		array $mime_filter,
-		string $modified_since,
-		int $depth,
-		array &$visited,
-		array &$logs
-	) {
-		if ( isset( $visited[ $folder_id ] ) ) {
-			return array();
-		}
-		$visited[ $folder_id ] = true;
-
-		if ( $depth > self::MAX_RECURSION ) {
-			$logs[] = array(
-				'level'   => 'warning',
-				'message' => 'GoogleDrive: Max recursion depth reached, skipping deeper folders.',
-				'data'    => array( 'folder_id' => $folder_id, 'depth' => $depth ),
-			);
-			return array();
-		}
-
-		$query_parts = array(
-			sprintf( "'%s' in parents", $this->escape_query_literal( $folder_id ) ),
-			'trashed = false',
-		);
-
-		if ( '' !== $modified_since ) {
-			$query_parts[] = sprintf( "modifiedTime > '%s'", $this->escape_query_literal( $modified_since ) );
-		}
-
-		// Note: MIME filter is applied client-side as well so that
-		// subfolders (mimeType = application/vnd.google-apps.folder) are
-		// not excluded by a too-narrow server-side filter when recursive
-		// is on.
-
-		$page_token = '';
-		$collected  = array();
-
-		do {
-			$params = array(
-				'q'                         => implode( ' and ', $query_parts ),
-				'pageSize'                  => self::PAGE_SIZE,
-				'fields'                    => self::LIST_FIELDS,
-				'supportsAllDrives'         => 'true',
-				'includeItemsFromAllDrives' => 'true',
-			);
-			if ( '' !== $page_token ) {
-				$params['pageToken'] = $page_token;
-			}
-
-			$response = $this->drive_get( self::API_BASE . '/files', $params, $access_token, $logs );
-			if ( is_wp_error( $response ) ) {
-				return $response;
-			}
-
-			$files      = $response['files'] ?? array();
-			$page_token = $response['nextPageToken'] ?? '';
-
-			foreach ( $files as $file ) {
-				if ( $this->is_folder( $file ) ) {
-					if ( $recursive ) {
-						$nested = $this->list_folder_recursive(
-							$file['id'] ?? '',
-							$access_token,
-							$recursive,
-							$mime_filter,
-							$modified_since,
-							$depth + 1,
-							$visited,
-							$logs
-						);
-						if ( is_wp_error( $nested ) ) {
-							return $nested;
-						}
-						foreach ( $nested as $nested_file ) {
-							$collected[] = $nested_file;
-						}
-					}
-					continue;
-				}
-
-				if ( ! empty( $mime_filter ) ) {
-					$file_mime = $file['mimeType'] ?? '';
-					if ( ! in_array( $file_mime, $mime_filter, true ) ) {
-						continue;
-					}
-				}
-
-				$collected[] = $file;
-			}
-		} while ( '' !== $page_token );
-
-		return $collected;
-	}
-
-	/**
 	 * Convert a raw Drive file record into a payload (text, local_path, or skipped).
 	 *
-	 * @param array  $file            Raw Drive file metadata.
-	 * @param string $access_token    OAuth bearer.
-	 * @param array  $download_target Pipeline/flow context for binary downloads. Empty = skip download.
-	 * @param array  $logs            Logs reference.
+	 * @param array             $file            Raw Drive file metadata.
+	 * @param GoogleDriveClient $client          Drive client (pre-authenticated).
+	 * @param array             $download_target Pipeline/flow context. Empty = skip binary download.
+	 * @param array             $logs            Logs reference.
 	 * @return array Payload descriptor.
 	 */
-	private function materialize_file( array $file, string $access_token, array $download_target, array &$logs ): array {
+	private function materialize_file( array $file, GoogleDriveClient $client, array $download_target, array &$logs ): array {
 		$mime    = $file['mimeType'] ?? '';
 		$file_id = $file['id'] ?? '';
 		$name    = $file['name'] ?? '';
 
-		// Native Google types use files.export; everything else uses files.get?alt=media.
-		$export_candidates = $this->export_candidates_for_native( $mime );
+		$export_candidates = GoogleDriveClient::export_candidates_for_native( $mime );
 
 		if ( null !== $export_candidates ) {
 			if ( empty( $export_candidates ) ) {
-				// Native but not exportable (Forms, Sites, etc.).
 				return array(
 					'skipped' => true,
 					'reason'  => 'Native Google file type is not exportable as text.',
@@ -351,8 +198,7 @@ class FetchGoogleDriveAbility {
 
 			$last_error = null;
 			foreach ( $export_candidates as $candidate_mime ) {
-				$url      = self::API_BASE . '/files/' . rawurlencode( $file_id ) . '/export';
-				$response = $this->drive_get_raw( $url, array( 'mimeType' => $candidate_mime ), $access_token, $logs );
+				$response = $client->export_native( $file_id, $candidate_mime, $logs );
 
 				if ( ! is_wp_error( $response ) ) {
 					return array(
@@ -386,7 +232,7 @@ class FetchGoogleDriveAbility {
 		}
 
 		// Binary file. Stream to disk if we have somewhere to put it.
-		if ( empty( $download_target['pipeline_id'] ) && empty( $download_target['flow_id'] ) ) {
+		if ( empty( $download_target['pipeline_id'] ) || empty( $download_target['flow_id'] ) ) {
 			$logs[] = array(
 				'level'   => 'debug',
 				'message' => 'GoogleDrive: No download target provided, listing binary file without downloading.',
@@ -397,7 +243,28 @@ class FetchGoogleDriveAbility {
 			);
 		}
 
-		$saved = $this->download_binary( $file_id, $name, $access_token, $download_target, $logs );
+		$pipeline_id = (int) $download_target['pipeline_id'];
+		$flow_id     = (int) $download_target['flow_id'];
+
+		$dir_manager = new DirectoryManager();
+		$directory   = $dir_manager->get_flow_files_directory( $pipeline_id, $flow_id );
+
+		if ( ! $dir_manager->ensure_directory_exists( $directory ) ) {
+			$logs[] = array(
+				'level'   => 'error',
+				'message' => 'GoogleDrive: Failed to create flow files directory.',
+				'data'    => array( 'directory' => $directory ),
+			);
+			return array(
+				'skipped' => true,
+				'reason'  => 'Failed to create flow files directory.',
+			);
+		}
+
+		$safe_name   = sanitize_file_name( $name ?: ( 'drive-' . $file_id ) );
+		$destination = trailingslashit( $directory ) . wp_unique_filename( $directory, $safe_name );
+
+		$saved = $client->download_binary( $file_id, $destination, $logs );
 		if ( is_wp_error( $saved ) ) {
 			$logs[] = array(
 				'level'   => 'error',
@@ -410,305 +277,11 @@ class FetchGoogleDriveAbility {
 			);
 		}
 
-		return $saved;
-	}
-
-	/**
-	 * Map a native Google MIME type to its preferred export target MIMEs.
-	 *
-	 * Returns an ORDERED list of MIME candidates to try via files.export.
-	 * The first candidate that the Drive API accepts wins. This lets us
-	 * prefer the LLM-friendlier text/markdown for Google Docs while
-	 * falling back to text/plain if Drive rejects it (older accounts or
-	 * documents with features the markdown exporter doesn't handle).
-	 *
-	 * Returns:
-	 * - null if the file is NOT a native Google type (caller falls back
-	 *   to files.get?alt=media for binary download).
-	 * - [] (empty array) if it IS a native type but not exportable as
-	 *   text (Forms, Sites, etc.).
-	 * - non-empty list of MIME candidates otherwise.
-	 *
-	 * @param string $mime Source MIME.
-	 * @return string[]|null
-	 */
-	private function export_candidates_for_native( string $mime ): ?array {
-		switch ( $mime ) {
-			case 'application/vnd.google-apps.document':
-				// Markdown is the LLM-preferred format. Drive started
-				// supporting it as a Docs export target in 2024 but older
-				// project credentials still 4xx on it — fall back to
-				// text/plain so the export still succeeds.
-				return array( 'text/markdown', 'text/plain' );
-			case 'application/vnd.google-apps.spreadsheet':
-				return array( 'text/csv' );
-			case 'application/vnd.google-apps.presentation':
-				return array( 'text/plain' );
-			case 'application/vnd.google-apps.form':
-			case 'application/vnd.google-apps.site':
-			case 'application/vnd.google-apps.drawing':
-			case 'application/vnd.google-apps.map':
-			case 'application/vnd.google-apps.shortcut':
-			case 'application/vnd.google-apps.folder':
-				return array();
-		}
-
-		if ( 0 === strpos( $mime, 'application/vnd.google-apps.' ) ) {
-			// Unknown native type — refuse rather than guess.
-			return array();
-		}
-
-		return null;
-	}
-
-	private function is_folder( array $file ): bool {
-		return ( $file['mimeType'] ?? '' ) === 'application/vnd.google-apps.folder';
-	}
-
-	/**
-	 * Download a binary Drive file into the flow-scoped uploads dir.
-	 *
-	 * Authenticated downloads need a bearer header, so we cannot reuse
-	 * WordPress's download_url() helper. We stream via wp_remote_get to
-	 * a temp file, then move into the flow directory. This is local-only
-	 * pending a generic authenticated-download helper in core (see
-	 * Extra-Chill/data-machine issue tracking TBD).
-	 *
-	 * @return array|\WP_Error
-	 */
-	private function download_binary( string $file_id, string $name, string $access_token, array $download_target, array &$logs ) {
-		$pipeline_id = $download_target['pipeline_id'] ?? null;
-		$flow_id     = $download_target['flow_id'] ?? null;
-		if ( null === $pipeline_id || null === $flow_id ) {
-			return new \WP_Error( 'googledrive_no_target', 'Download target missing pipeline_id/flow_id.' );
-		}
-
-		$dir_manager = new DirectoryManager();
-		$directory   = $dir_manager->get_flow_files_directory( $pipeline_id, $flow_id );
-
-		if ( ! $dir_manager->ensure_directory_exists( $directory ) ) {
-			return new \WP_Error( 'googledrive_dir_failed', 'Failed to create flow files directory.' );
-		}
-
-		$url      = self::API_BASE . '/files/' . rawurlencode( $file_id );
-		$response = wp_remote_get(
-			add_query_arg(
-				array(
-					'alt'               => 'media',
-					'supportsAllDrives' => 'true',
-				),
-				$url
-			),
-			array(
-				'timeout' => 60,
-				'headers' => array(
-					'Authorization' => 'Bearer ' . $access_token,
-				),
-				// Avoid loading the whole binary into PHP memory when possible.
-				'stream'   => true,
-				'filename' => trailingslashit( get_temp_dir() ) . wp_unique_filename( get_temp_dir(), 'gdrive-' . $file_id . '-' . sanitize_file_name( $name ) ),
-			)
-		);
-
-		if ( is_wp_error( $response ) ) {
-			return $response;
-		}
-
-		$status_code = wp_remote_retrieve_response_code( $response );
-		$temp_path   = $response['filename'] ?? '';
-
-		if ( 200 !== (int) $status_code ) {
-			if ( $temp_path && file_exists( $temp_path ) ) {
-				wp_delete_file( $temp_path );
-			}
-			return $this->error_from_status( $status_code, $response );
-		}
-
-		if ( '' === $temp_path || ! file_exists( $temp_path ) ) {
-			return new \WP_Error( 'googledrive_stream_failed', 'Drive returned 200 but temp file is missing.' );
-		}
-
-		$safe_name   = sanitize_file_name( $name ?: ( 'drive-' . $file_id ) );
-		$destination = trailingslashit( $directory ) . wp_unique_filename( $directory, $safe_name );
-
-		$fs = FilesystemHelper::get();
-		if ( $fs ) {
-			$copied = $fs->copy( $temp_path, $destination, true );
-		} else {
-			// phpcs:ignore WordPress.WP.AlternativeFunctions.rename_rename -- Filesystem API unavailable, fall back to native move.
-			$copied = @rename( $temp_path, $destination );
-		}
-
-		if ( file_exists( $temp_path ) ) {
-			wp_delete_file( $temp_path );
-		}
-
-		if ( ! $copied || ! file_exists( $destination ) ) {
-			return new \WP_Error( 'googledrive_move_failed', 'Failed to move downloaded file into flow directory.' );
-		}
-
-		$logs[] = array(
-			'level'   => 'debug',
-			'message' => 'GoogleDrive: Binary file downloaded.',
-			'data'    => array( 'file_id' => $file_id, 'destination' => $destination ),
-		);
-
 		return array(
-			'local_path' => $destination,
-			'filename'   => basename( $destination ),
-			'size'       => filesize( $destination ),
+			'local_path' => $saved['local_path'],
+			'filename'   => basename( $saved['local_path'] ),
+			'size'       => $saved['size'],
 		);
-	}
-
-	/**
-	 * GET a JSON-returning Drive API endpoint.
-	 *
-	 * @return array|\WP_Error Decoded JSON on success.
-	 */
-	private function drive_get( string $url, array $params, string $access_token, array &$logs ) {
-		$full_url = add_query_arg( $params, $url );
-		$response = wp_remote_get(
-			$full_url,
-			array(
-				'timeout' => 30,
-				'headers' => array(
-					'Authorization' => 'Bearer ' . $access_token,
-					'Accept'        => 'application/json',
-				),
-			)
-		);
-
-		if ( is_wp_error( $response ) ) {
-			return $response;
-		}
-
-		$status_code = (int) wp_remote_retrieve_response_code( $response );
-		$body        = wp_remote_retrieve_body( $response );
-
-		if ( 200 !== $status_code ) {
-			return $this->error_from_status( $status_code, $response, $body );
-		}
-
-		$decoded = json_decode( $body, true );
-		if ( ! is_array( $decoded ) ) {
-			return new \WP_Error( 'googledrive_decode_failed', 'Drive returned an unparseable response body.' );
-		}
-
-		return $decoded;
-	}
-
-	/**
-	 * GET a Drive API endpoint that returns raw (non-JSON) content.
-	 *
-	 * Used for files.export.
-	 *
-	 * @return string|\WP_Error Raw body on success.
-	 */
-	private function drive_get_raw( string $url, array $params, string $access_token, array &$logs ) {
-		$full_url = add_query_arg( $params, $url );
-		$response = wp_remote_get(
-			$full_url,
-			array(
-				'timeout' => 60,
-				'headers' => array(
-					'Authorization' => 'Bearer ' . $access_token,
-				),
-			)
-		);
-
-		if ( is_wp_error( $response ) ) {
-			return $response;
-		}
-
-		$status_code = (int) wp_remote_retrieve_response_code( $response );
-		$body        = (string) wp_remote_retrieve_body( $response );
-
-		if ( 200 !== $status_code ) {
-			return $this->error_from_status( $status_code, $response, $body );
-		}
-
-		return $body;
-	}
-
-	/**
-	 * Translate a non-200 Drive response into a WP_Error.
-	 *
-	 * Distinguishes scope-missing (403 with insufficient scope reason),
-	 * not-found (404), and rate-limited (429, with respect for
-	 * Retry-After) errors.
-	 *
-	 * @param int          $status_code HTTP status.
-	 * @param array|object $response    wp_remote_* response.
-	 * @param string|null  $body        Body string if already retrieved.
-	 * @return \WP_Error
-	 */
-	private function error_from_status( int $status_code, $response, ?string $body = null ): \WP_Error {
-		if ( null === $body ) {
-			$body = (string) wp_remote_retrieve_body( $response );
-		}
-
-		$decoded = json_decode( $body, true );
-		$message = '';
-		$reason  = '';
-		if ( is_array( $decoded ) && isset( $decoded['error'] ) ) {
-			$message = (string) ( $decoded['error']['message'] ?? '' );
-			$errors  = $decoded['error']['errors'] ?? array();
-			if ( is_array( $errors ) && ! empty( $errors[0]['reason'] ) ) {
-				$reason = (string) $errors[0]['reason'];
-			}
-		}
-
-		if ( 401 === $status_code ) {
-			return new \WP_Error( 'googledrive_unauthorized', 'Google Drive returned 401 Unauthorized. Re-authenticate the Google integration.' );
-		}
-
-		if ( 403 === $status_code ) {
-			if ( in_array( $reason, array( 'insufficientPermissions', 'insufficientScopes', 'forbidden' ), true ) && false !== stripos( $message . ' ' . $reason, 'scope' ) ) {
-				return new \WP_Error(
-					'googledrive_scope_missing',
-					sprintf(
-						'Google Drive denied the request because the OAuth token lacks the required scope (%s). Disconnect and reconnect the Google integration to grant Drive access.',
-						self::REQUIRED_SCOPE
-					)
-				);
-			}
-			return new \WP_Error(
-				'googledrive_forbidden',
-				$message ?: 'Google Drive denied access to this resource. The authenticated user may not have permission to view the folder.'
-			);
-		}
-
-		if ( 404 === $status_code ) {
-			return new \WP_Error( 'googledrive_not_found', $message ?: 'Drive resource not found.' );
-		}
-
-		if ( 429 === $status_code ) {
-			$retry_after = (int) wp_remote_retrieve_header( $response, 'retry-after' );
-			if ( $retry_after > 0 ) {
-				$retry_after = min( $retry_after, self::RETRY_AFTER_CAP );
-				return new \WP_Error(
-					'googledrive_rate_limited',
-					sprintf( 'Google Drive rate limit hit. Retry after %d seconds.', $retry_after ),
-					array( 'retry_after' => $retry_after )
-				);
-			}
-			return new \WP_Error( 'googledrive_rate_limited', 'Google Drive rate limit hit.' );
-		}
-
-		return new \WP_Error(
-			'googledrive_http_error',
-			sprintf( 'Google Drive API returned HTTP %d: %s', $status_code, $message ?: 'Unknown error' )
-		);
-	}
-
-	/**
-	 * Escape a literal for inclusion inside a Drive `q=` query string.
-	 *
-	 * Drive's query language quotes literals with single quotes and
-	 * escapes single quotes and backslashes inside them.
-	 */
-	private function escape_query_literal( string $value ): string {
-		return str_replace( array( '\\', "'" ), array( '\\\\', "\\'" ), $value );
 	}
 
 	private function fail( string $message, array $logs ): array {
@@ -718,14 +291,5 @@ class FetchGoogleDriveAbility {
 			'error'   => $message,
 			'logs'    => $logs,
 		);
-	}
-
-	/**
-	 * @return \DataMachineBusiness\OAuth\Providers\GoogleAuth|null
-	 */
-	private static function get_auth_provider(): ?\DataMachineBusiness\OAuth\Providers\GoogleAuth {
-		$providers = apply_filters( 'datamachine_auth_providers', array() );
-		$provider  = $providers['google'] ?? null;
-		return $provider instanceof \DataMachineBusiness\OAuth\Providers\GoogleAuth ? $provider : null;
 	}
 }
